opto_tagging_protocol_extStim.py
This file allows you to do an opto-tagging protocol.

It requires the Labview GUI for waveforms, and the External Arduino to send the actual TTL to labview.

What this file does allow you to do is simply set the time between each TTL in seconds ( interval=2,) and set how long the duration will be in seconds (duration=120)

in this example     optotagging_protocol(port="COM5", interval=2, duration=120)
you will get 60 pulses sent, each 2 seconds apart for 120 seconds.